+++ addOns/selenium/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added stealth mode to make the browser appear to be controlled by a user (Chrome only).
+
 ### Changed
 - Update Selenium to version 4.31.0.